+++ auth/src/main/java/kernel360/ckt/auth/application/AuthService.java
@@ -3,7 +3,7 @@ package kernel360.ckt.auth.application;
 import kernel360.ckt.auth.application.command.LoginCommand;
 import kernel360.ckt.auth.application.command.ReissueCommand;
 import kernel360.ckt.auth.config.JwtTokenProvider;
-import kernel360.ckt.auth.domain.RefreshTokenStatus;
+import kernel360.ckt.core.domain.enums.RefreshTokenStatus;
 import kernel360.ckt.auth.infra.CompanyReadJpaRepository;
 import kernel360.ckt.auth.infra.RefreshTokenEntity;
 import kernel360.ckt.auth.infra.RefreshTokenJpaRepository;
